refactor: resolver 분리 및 dto 네이밍 변경
