+++ src/main/java/antigravity/common/resolver/UserId.java
@@ -1,4 +1,4 @@
-package antigravity.config;
+package antigravity.common.resolver;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
