+++ src/main/java/antigravity/config/WebMvcConfig.java
@@ -6,6 +6,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+import antigravity.common.resolver.UserIdArgumentResolver;
+
 @Configuration
 public class WebMvcConfig implements WebMvcConfigurer {
 	private final UserIdArgumentResolver userIdArgumentResolver;
