--- src/main/java/antigravity/repository/custom/CustomProductRepository.java
@@ -1,8 +1,0 @@
-package antigravity.repository.custom;
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CustomProductRepository {
-
-}
