--- src/main/java/antigravity/repository/custom/CustomProductRepositoryImpl.java
@@ -1,11 +1,0 @@
-package antigravity.repository.custom;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-public class CustomProductRepositoryImpl implements CustomProductRepository {
-	private final JPAQueryFactory jpaQueryFactory;
-
-}
